requirements.txt
flask~=3.0.3
scikit-learn~=1.4.2
loguru~=0.5.3
rich~=13.7.1
transformers~=4.41.1
datasets~=2.19.1
pandas~=2.2.2
tqdm~=4.66.4
joblib~=1.4.0
xgboost~=2.0.3
seaborn~=0.12.2
flasgger~=0.9.7.1
streamlit~=1.35.0
python-dotenv~=1.0.1
openai~=1.30.3